assessment new properties; referral sources list; add assessment with appointment day;

+++ src/Web/PsychologicalServices.Models/Referrals/ReferralSource.cs
@@ -14,7 +14,7 @@ namespace PsychologicalServices.Models.Referrals
 
         public int LargeFileSize { get; set; }
 
-        public decimal LargeFileFeeAmount { get; set; }
+        public int LargeFileFeeAmount { get; set; }
 
         public bool IsActive { get; set; }
